% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSpecimenColour.R
\name{plot.colours}
\alias{plot.colours}
\title{Helper function: do not export. Plot RGB and CYM colours}
\usage{
\method{plot}{colours}(
  colour.data,
  colour.data.from.all.colour.models,
  population,
  output.folder
)
}
\arguments{
\item{colour.data}{Data frame containing sample information including Accession Number (or any unique identifier for each sample), sample grouping(s), and a column each for R, G and B values}

\item{colour.data.from.all.colour.models}{Results from add.colour.models.from.RGB()}

\item{population}{Name of column with population assignment (e.g. subspecies, general locality etc.)}

\item{output.folder}{Path to folder where image file will be exported into. If left blank, the file will be created in present working directory. If a folder with the name designated here doesn't exist in the present workign directory, it will be created.}
}
\description{
Helper function: do not export. Plot RGB and CYM colours
}
